_parent_stack reliability improvements
I'm seeing a situation whereby _parent_stack may not be set in some cases.
This causes `traceback.format_list(new_stack)` to error because it is passed
a null value.

I'm therefore changing how we check for the presense of _parent_stack, and
also moving the setting of _parent_stack a little higher up.

_parent_stack isn't critical at any rate, so stricter check
regarding its use in the name of avoiding errors seems sane.